Add files via upload
In Task 9, I have created a calculator that computes basic arithmetic problems. It also records all the user's computation and writes them into a .txt file and provides the user the opportunity to view this record. I have ensured my program is robust and handles possible user errors to ensure it does not crash.